Revert "Refactor codebase"
