--- controllers/__init__.py
@@ -1,12 +1,0 @@
-"""
-Application controllers for managing interactions between UI and core logic.
-"""
-from .input_handler import UserInputHandler
-from .view_controller import ViewController
-from .annotation_controller import AnnotationController
-
-__all__ = [
-    'UserInputHandler',
-    'ViewController', 
-    'AnnotationController'
-]
