--- core/__init__.py
@@ -1,6 +1,0 @@
-"""
-Core business logic for Inkshade PDF Reader.
-"""
-from .annotations import AnnotationManager, Annotation, AnnotationType
-
-__all__ = ['AnnotationManager', 'Annotation', 'AnnotationType']
