--- core/document/__init__.py
@@ -1,7 +1,0 @@
-"""
-PDF document handling and manipulation.
-"""
-from .pdf_reader import PDFDocumentReader
-from .pdf_exporter import PDFExporter
-
-__all__ = ['PDFDocumentReader', 'PDFExporter']
