--- core/export/__init__.py
@@ -1,6 +1,0 @@
-"""
-Export operations for PDF documents.
-"""
-from .export_worker import ExportWorker
-
-__all__ = ['ExportWorker']
