+++ helpers/locate_resources.py
@@ -1,0 +1,7 @@
+import os
+import sys
+
+def get_resource_path(relative_path):
+    if hasattr(sys, '_MEIPASS'):
+        return os.path.join(sys._MEIPASS, relative_path)
+    return os.path.join(os.path.abspath('.'), relative_path)
