+++ styles/__init__.py
@@ -1,6 +1,1 @@
-"""
-Application styling and themes.
-"""
-from .theme_manager import ThemeColors, ThemeManager
-
-__all__ = ['ThemeColors', 'ThemeManager']
+from .styles import apply_style
