+++ ui/__init__.py
@@ -1,4 +1,1 @@
-"""
-User interface components for Inkshade PDF Reader.
-"""
-# Main UI components are in subpackages
+from .main_window import MainWindow
