+++ ui/search_bar.py
