--- ui/toolbars/__init__.py
@@ -1,8 +1,0 @@
-"""
-Toolbar components for PDF operations.
-"""
-from .annotation_toolbar import AnnotationToolbar
-from .drawing_toolbar import DrawingToolbar
-from .search_bar import SearchBar
-
-__all__ = ['AnnotationToolbar', 'DrawingToolbar', 'SearchBar']
