--- ui/windows/__init__.py
@@ -1,6 +1,0 @@
-"""
-Main application windows.
-"""
-from .main_window import MainWindow
-
-__all__ = ['MainWindow']
