--- utils/__init__.py
@@ -1,6 +1,0 @@
-"""
-Utility functions and helpers.
-"""
-from .resource_loader import get_resource_path, get_icon_path
-
-__all__ = ['get_resource_path', 'get_icon_path']
